1 - rename ircam_oauth2_provider to allauth_ircam. 2 - adaptation to Mezzo
