+++ allauth_ircam/urls.py
@@ -2,8 +2,8 @@ from allauth.socialaccount.providers.oauth2.urls import default_urlpatterns,url
 from .provider import IrcamAuthProvider
 from django.conf import settings
 
-try:
-    if settings.OAUTH2_IRCAM == True:
+if getattr(settings, 'OAUTH2_IRCAM', False):
+    if settings.OAUTH2_IRCAM:
         urlpatterns = default_urlpatterns(IrcamAuthProvider)
-except AttributeError:
+else:
     urlpatterns = default_urlpatterns(IrcamAuthProvider)
